requirements-frozen-3.4.txt
coverage==4.0
decorator==4.0.4
Django==1.8.5
django-autocomplete-light==2.2.10
django-daterange-filter==1.1.1
django-debug-panel==0.8.1
django-debug-toolbar==1.4
django-extensions==1.5.7
django-filebrowser==3.6.1
django-filter==0.11.0
django-grappelli==2.7.1
django-grappelli-filters==0.2.1
django-gunicorn==0.1.1
django-mptt==0.7.4
django-mptt-admin==0.3.0
django-nose==1.4.2
django-sphinxql==1.0.0
djangorestframework==3.2.4
ecdsa==0.13
Fabric==1.10.2
gunicorn==19.3.0
ipython==4.0.0
ipython-genutils==0.1.0
Markdown==2.6.2
mock==1.3.0
model-mommy==1.2.5
nose==1.3.7
paramiko==1.15.3
path.py==8.1.2
pbr==1.8.1
pexpect==4.0.1
pickleshare==0.5
Pillow==3.0.0
psycopg2==2.6.1
ptyprocess==0.5
pycrypto==2.6.1
PyMySQL==0.6.7
simplegeneric==0.8.1
six==1.10.0
sqlparse==0.1.16
traitlets==4.0.0
Unidecode==0.4.18
